Resolve issue where cmacc variables were being written twice.
This was a two part problem.  We add fields from the Edit and Complete
process, then we only add fields from the .dot file, causing duplicates.
Now checks to see if the field exists and if found updates it.

Also if the field does not have a value, it is not written to the .md